Update serialproto.c
The program now reads the first 4 bytes of data coming from the host and stores the data in a header. The header is parsed and a buffer the size of length data extracted from the header is created. This means that there is now have a buffer big enough to grab the entire packet so handling data becomes trivial.